finish implementing game over screen, allow key input for setting name, enter to submit + restart, escape to restart (without submitting score)

+++ core/src/dev/therealdan/tetris/game/GameInstance.java
@@ -10,7 +10,7 @@ public class GameInstance {
     public HashSet<Tetrimino> tetriminos = new HashSet<>();
     public HashSet<Square> squares = new HashSet<>();
     public Deque<Tetrimino.Type> tetriminoQueue = new ArrayDeque<>();
-    public double score = 0;
+    public int score = 0;
 
     private long fallInterval = 1000;
